Create app.ts and import from index
This will make it possible to run unit tests on app.ts
